Revert "further extracted logic into a helper class for user states"
This reverts commit 5674f34ef72bb75ce677492639902d57fc204642.
