--- Explorer/Assets/DCL/Chat/ChatControllerUserStateHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2867c03791cbe5f42b8ae8bc9dd468ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
