+++ Explorer/Assets/DCL/Chat/IChatController.cs
@@ -1,5 +1,3 @@
-using DCL.UI;
-
 namespace DCL.Chat
 {
     public interface IChatController
@@ -7,7 +5,5 @@ namespace DCL.Chat
         public string IslandRoomSid { get; }
         public string PreviousRoomSid { get; set; }
         public bool TryGetView(out ChatView view);
-        void SetInputWithUserState(ChatUserStateUpdater.ChatUserState state);
-        void UpdateConversationToolbarStatusIcon(string userId, OnlineStatus status);
     }
 }
